Rename ComponentRegistry to ProviderRegistry
